feat: extract sfc-shared-utils and sfc-compiler
