+++ packages/mp-runtime/package.json
@@ -8,7 +8,7 @@
     "build": "cross-env NODE_ENV=production webpack --config webpack.config.js"
   },
   "dependencies": {
-    "my-compat": "latest"
+    "my-compact": "latest"
   },
   "devDependencies": {
     "babel-loader": "^7.1.4",
