+++ packages/sfc-compiler/codegen/createRenderFn.js
@@ -1,12 +1,12 @@
-const { declearTags } = require('../../transform/declear');
-const { uniqueInstanceID, makeMap } = require('../../shared/utils');
+const { declearTags } = require('../declear');
+const { uniqueInstanceID, makeMap } = require('sfc-shared-utils');
 const injectThisScope = require('./injectThisScope');
 
 // helpers
 const declarationName = `$_${uniqueInstanceID}_declaration`;
 const helpersFns = '_c,_o,_n,_s,_l,_t,_q,_i,_m,_f,_k,_b,_v,_e,_u,_g,_cx';
 
-module.exports = function(render, tagHelperMap) {
+module.exports = function (render, tagHelperMap) {
   const declearedTags = declearTags(tagHelperMap, '$t');
   const existRenderHelpers = makeMap(
     helpersFns + ',' + Object.values(tagHelperMap).join(',')
