+++ packages/sfc-compiler/codegen/injectThisScope.js
@@ -1,5 +1,5 @@
-const { makeMap } = require('../../shared/utils');
-const { uniqueInstanceID } = require('../../shared/utils');
+const { makeMap } = require('sfc-shared-utils');
+const { uniqueInstanceID } = require('sfc-shared-utils');
 const babylon = require('babylon');
 const traverse = require('babel-traverse').default;
 const t = require('babel-types');
@@ -19,7 +19,7 @@ const isPreveredGlobalObject = makeMap(
  * @param {Function} existsScope 白名单
  * @param {String} prefix 前缀
  */
-module.exports = function(code, existsScope = () => false, prefix = 'this') {
+module.exports = function (code, existsScope = () => false, prefix = 'this') {
   let ast;
   try {
     ast = babylon.parse(code, {
@@ -111,7 +111,7 @@ module.exports = function(code, existsScope = () => false, prefix = 'this') {
     .map(
       name =>
         `var ${name} = ${prefix}${
-          isValidIdentifier(name) ? '.' + name : "['" + name + "']"
+        isValidIdentifier(name) ? '.' + name : "['" + name + "']"
         };`
     )
     .join('\n');
