+++ packages/sfc-compiler/declear.js
@@ -4,7 +4,7 @@ const traverse = require('babel-traverse').default;
 const generate = require('babel-generator').default;
 const t = require('babel-types');
 const uppercamelcase = require('uppercamelcase');
-const { uniqueInstanceID } = require('../shared/utils');
+const { uniqueInstanceID } = require('sfc-shared-utils');
 
 /**
  * kv: {
@@ -14,7 +14,7 @@ const { uniqueInstanceID } = require('../shared/utils');
  * bar is identifier
  * name: is variable name
  */
-exports.declearObject = function(kv, name) {
+exports.declearObject = function (kv, name) {
   const properties = Object.keys(kv).map(key => {
     const value = kv[key];
     return t.objectProperty(
@@ -31,7 +31,7 @@ exports.declearObject = function(kv, name) {
   return generate(ast).code;
 };
 
-exports.declearVariables = function(kv) {
+exports.declearVariables = function (kv) {
   const declearations = Object.keys(kv).map(key => {
     const value = kv[key];
 
@@ -54,7 +54,7 @@ exports.declearVariables = function(kv) {
  * @param {Object} tags
  * @param {String} parent
  */
-exports.declearTags = function(tags, parentName) {
+exports.declearTags = function (tags, parentName) {
   const declearators = Object.keys(tags).map(sourceTagName => {
     const targetName = tags[sourceTagName];
