+++ packages/sfc-compiler/utils/style.js
@@ -1,10 +1,10 @@
-const { cached, extend, toObject } = require('../../shared/utils');
+const { cached, extend, toObject } = require('sfc-shared-utils');
 
-const parseStyleText = cached(function(cssText) {
+const parseStyleText = cached(function (cssText) {
   const res = {};
   const listDelimiter = /;(?![^(]*\))/g;
   const propertyDelimiter = /:(.+)/;
-  cssText.split(listDelimiter).forEach(function(item) {
+  cssText.split(listDelimiter).forEach(function (item) {
     if (item) {
       var tmp = item.split(propertyDelimiter);
       tmp.length > 1 && (res[tmp[0].trim()] = tmp[1].trim());
